Created edit.php using CakePHP bake script
Used the CakePHP bake script to create a CRUD model of the Workorders2021 spreadsheet and transferred all of the data over to the database so that it could be edited within the app itself, instead of relying on the Google spreadsheet and App script.  This is a more streamlined approach.