Created index.php as main directory for all data
Used the CakePHP bake script to create an index of the whole Workorders2021 data set so that it can be browsed, filtered, searched through and act as the main directory screen to begin searching and requesting data from.  The index view contains the action buttons for the other baked scripts add.php, edit.php, and view.php.  You can also delete a record from the index as well by using the "Delete" action button.