Created login.php for user authentication
The login.php template was baked using the CakePHP bake script and was created for the Users table and the Workorders2021 table so that these 2 tables can only be viewed by a authorized user that must login and have the correct permissioin to gain access to the data in them.  Users will be able to create an account on the main registration page and permissions will be determined during the registration process, depending on if you are a Service Team Tier 1 or Tier 2 Member or upper management/administration.