Created login.php for user authentication
Users will need to be logged in to gain access to certain tables like Workorders2021 and Users etc.  Use of the login.php will also be in conjunction with the register.php or add.php template files in order to create a place for users to create an account and have them go to the login.php for authentication and validation of their user information for use within the app.